% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalMWStat.R
\name{MwStatistics}
\alias{MwStatistics}
\title{MwStatistics}
\usage{
MwStatistics(dat = NULL, TimeVarCol = 1, ColA = 2, ColB = 3,
win = NULL, stat = NULL, overlap = 0, genplot = TRUE, genDataFrame = FALSE)
}
\arguments{
\item{dat}{Name of dataset.}

\item{TimeVarCol}{Column number of time variable (EG: Depth or Time)}

\item{ColA}{Column number of first variable}

\item{ColB}{Column number of second variable}

\item{win}{Window Size}

\item{stat}{Calculated Statistic, Requires Quotes}

\item{overlap}{Overlap between windows}

\item{genplot}{If you want a plot to be generated}

\item{genDataFrame}{If you want a dataframe to be generated}
}
\value{
Plot and/or Dataframe
}
\description{
This function allows you to calculate a variety of statistics in a time series using a moving window.
}
\examples{
MwStatistics(dat = data.frame(replicate(10,sample(0:10,1000,rep=TRUE))),
TimeVarCol = 1, ColA = 2, ColB = 3, win = 2, stat = "Correlation",
genplot = TRUE, genDataFrame = FALSE, overlap = 0.5)
}
\author{
William Kopans -  \email{wkopans123@gmail.com}
}
